料金規則を施行日別に構成
