+++ docs/design-principles.md
@@ -57,8 +57,9 @@
 - `App.tsx`は現在表示する画面だけを持ち、旧画面をアーカイブ目的で残しません。
 - `domain/routes.ts`は両年版で共通して使う駅・路線・営業キロと、特定都区市内等を含む区間操作を持ちます。年版固有の賃率・料金・交換ポイントは置きません。
 - `domain/seasons.ts`は両年版で共通して使うシーズン区分を持ちます。
-- `domain/versions/2022.ts`は2022年版の運賃・料金・交換ポイントをまとめた計算器とします。
-- `domain/versions/2026.ts`は2026年版の運賃・料金・交換ポイントをまとめた計算器とします。
+- `domain/fare-calculation.ts`は、旅客営業規則の表を検索して運賃・料金を構成する処理と、施行日または確認基準日を明示した規則部品を持ちます。
+- `domain/versions/2022.ts`は2022年3月12日時点で有効な普通運賃・通常期基準額・3段階のシーズン規則・交換ポイントを明示的に選ぶ計算器とします。
+- `domain/versions/2026.ts`は2026年3月14日時点で有効な普通運賃・通常期基準額・4段階のシーズン規則・交換ポイント・特別車両料金を明示的に選ぶ計算器とします。
 - `domain/quote.ts`で年版ごとの計算器を選択・構成し、Reactコンポーネントには計算式を置きません。
 - 画面内の「区間指定」と「ランキング」はページルーティングではなくタブとして扱い、ホスティングサービス固有のrewriteを必要としない構成にします。
 
@@ -76,6 +77,10 @@
 特定都区市内・東京山手線内は年版固有の金額ではなく共通の区間規則として扱い、適用後の中心駅から各年版の普通運賃を計算します。
 
 新幹線の駅間に直接定められた特定特急料金は、旅客営業規則別表第2号の三角表を原典とします。コード上のCSVはその値を検索するための表現であり、外部の乗換案内結果から作成しません。
+
+同じ金額が複数年版に現れることだけを理由に「共通値」として扱いません。通常期の新幹線特急料金表は、2022年3月12日時点の表と2026年3月14日時点の表が同額であることを確認した規則部品として参照します。法令上の施行開始日を確認できない場合は推測せず、確認基準日と同額を確認した期限を記録します。
+
+シーズン加減は通常期基準額と分離します。2022年3月12日版は閑散期・通常期・繁忙期の3段階、2026年3月14日版は2022年4月1日施行の最繁忙期を含む4段階を使用します。グリーン車・グランクラスについては、2023年4月1日施行のシーズン別特急料金を適用した後、通常期の指定席特急料金との差額530円を低減します。
 
 ### 東北新幹線と山形・秋田新幹線を通して利用する場合
